+++ __init__.py
@@ -1,7 +1,7 @@
 """netports"""
 
 from netports.interface4 import Interface4
-from netports.ip import IP_NAMES, IP_PORTS, iip, nip, sip
+from netports.ip import IP_NAMES, IP_NUMBERS, iip, nip, sip
 from netports.item import Item
 from netports.ports import inumbers, parse_range, snumbers
 from netports.range import Range